+++ src/bes/lims/subscribers/upgrade.py
@@ -4,6 +4,8 @@ from bes.lims import logger
 from bes.lims import PRODUCT_NAME
 from bes.lims.setuphandlers import setup_behaviors
 from bes.lims.setuphandlers import setup_catalogs
+from bes.lims.setuphandlers import setup_groups
+from bes.lims.setuphandlers import setup_roles
 from bes.lims.setuphandlers import setup_workflows
 from bika.lims.api import get_portal
 
@@ -19,6 +21,12 @@ def afterUpgradeStepHandler(event):  # noqa CamelCase
     profile = "profile-{0}:default".format(PRODUCT_NAME)
     setup.runImportStepFromProfile(profile, "actions")
     setup.runImportStepFromProfile(profile, "rolemap")
+
+    # Setup roles
+    setup_roles(portal)
+
+    # Setup groups
+    setup_groups(portal)
 
     # Setup catalogs
     setup_catalogs(portal)